Struct treediff::tools::Merger

pub struct Merger<K, V, BF, F> { /* fields omitted */ }

A Delegate which applies differences to a target object.

It wraps the target object and applies all calls by the diff algorithm to it, which changes it in some way.

Custom resolver functions can be provided to arbitrarily alter the way the merge is performed. This allows you, for example, to keep your own meta-data, or to implement custom conflict resolutions.

Examples

Please see the [tests][tests] for usage examples. [tests]: https://github.com/Byron/treediff-rs/blob/master/tests/merge.rs#L22

Implementations

impl<K, V, BF, F> Merger<K, V, BF, F>

pub fn into_inner(self) -> V

Consume the merger and return the contained target Value, which is the result of the merge operation.

pub fn filter_mut(&mut self) -> &mut BF

Returns a mutable borrow to the MutableFilter instance

pub fn filter(&self) -> &BF

Returns a borrow to the MutableFilter instance

impl<'a, V, BF, F> Merger<V::Key, V, BF, F> where
V: Mutable + 'a + Clone,
F: MutableFilter,
BF: BorrowMut<F>,

pub fn with_filter(v: V, f: BF) -> Self

Return a new Merger with the given initial value v and the filter f