Struct treediff::tools::Recorder

pub struct Recorder<'a, K, V: 'a> {
    pub calls: Vec<ChangeType<'a, K, V>>,
    /* private fields */
}

A Delegate to record all calls made to it.

It can be useful if you don’t want to implement your own custom delegate, but instead just want to quickly see a flat list of all results of the diff run.

Examples

Please see the tests for how to use this type.

Fields
calls: Vec<ChangeType<'a, K, V>>

A list of all calls the diff function made on us.